Initial commit
Like it works, but the code is not very flexible or user-firendly. There
are basic mechanisms to avoid double download and make it somewhat
multi-threaddable.

How does is work technically? DB ICEPortal is strucutred in a way that
alll of the page content is provided via a JSON API. By querying this
API we can extract all of the paths and a list of audiobooks (I don't
know if this list is complete). On the audiobooks page, we get paths to
all chapteras as individual audio files. If you call them up directly
your are forbidden. Therefore, this path needs to be send trough an
extra API to add an token to the path. Using this URL we can download it
directly to a file.

At the moment only audiobooks are supported and it's to be determined
if this script / I will support that. It's mostly due to a lack of time,
money and reasons to ride trains. Well, mostly money and some time.